requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
langchain==0.1.3
langgraph==0.0.20
python-dotenv==1.0.0
requests==2.31.0
pypdf==3.17.4
pandas==2.1.4
openpyxl==3.1.2
pytest==7.4.4
pytest-asyncio==0.23.3
jsonschema==4.21.1
reportlab==4.0.9
matplotlib==3.8.2
